Add scripts to upload repos to hf
