+++ bug_localization/configs/hf_data.yaml
@@ -2,13 +2,13 @@ data_path: /Users/Maria.Tigina/PycharmProjects/lca-baselines/data/lca-bug-locali
 repos_path: /Users/Maria.Tigina/PycharmProjects/lca-baselines/data/lca-bug-localization/repos
 test_data_ids: [
   # py
-  2328, 2759, 2760, 2763, 2762, 1929, 1932, 2409, 1890, 1885, 1889, 1888,
-  1887, 1886, 2076, 2075, 2067, 2060, 2062, 2064, 2065, 2074, 2066, 972,
-  971, 2294, 2287, 2295, 2289, 2476,
+  5119, 5406, 5396, 6918, 7211, 5838, 5987, 5883, 2701, 2698, 7617, 7098, 7446, 8131, 6491,
+  6497, 6498, 6487, 6839, 6667, 3414,
   # java
-  107,
+  759, 249, 6871, 225, 52, 71, 107, 106, 87, 145, 125, 153, 0, 104, 24, 26, 46, 84, 78, 80,
   # kotlin
-  108,
+  1442, 1441, 7671, 7672, 1920, 7749, 7752, 7753, 7755, 1450, 1552, 1553, 1472, 1614, 1415,
+  1419, 2012, 1403, 134, 1421, 1413, 1944,
   # mixed
-  2761
+  4108
 ]
