+++ bug_localization/data/hf/README.md
@@ -1,0 +1,10 @@
+# Scripts for data HF uploading
+
+### [upload_data.py](upload_data.py)
+Upload data to huggingface repo by categories to `dev` split
+
+### [split_data.py](split_data.py)
+Split into `dev` / `test` / `train` splits
+
+### [upload_repos.py](upload_repos.py)
+Upload repos for `test` split
